add construction of sky frames
A sky frame is the dominant response of the camera to the sky. It is
constructed by first subtracting a large-scale focal plane background model
(e.g., 2048x2048) from all frames to remove large-scale gradients, binning
pixels in each CCD (e.g., 64x64) and then averaging those.

This also includes code to generate secondary sky frames. These are
larger-scale (e.g., a large fraction of the size of a single CCD) than the
primary sky frame, and the intention is that these be used to correct for
chromatic response variations (e.g., as the sky color changes).

I'm not entirely happy with the sky frame subtractions I've gotten on HSC
yet (notably at the edge, where the vignetting kicks in), but this should
serve as a good foundation for future work.