add script to perform sky correction
We remove a large-scale background model, a scaled sky frame and then a
small-scale background model.  This does a decent job over most of the
field of view of HSC, but leaves annoying residuals near the edges.  More
work is necessary to get this ready for production use, but it should
serve as a good foundation for future work.

Currently, we read the 'calexp' and 'calexpBackground' datasets and then
write over the top (clobbering if not using a new rerun or output
data repo). While this is not really desirable, it saves us from having
different data products for downstream processing depending upon whether
sky correction was applied or not. It is expected that in production this
would be integrated with the singleFrameDriver and the write only done
once.